Added checking condiciton, if id != null

+++ src/main/java/com/assist/bookingjava/controllers/RedCompanyController.java
@@ -41,7 +41,11 @@ public class RedCompanyController {
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/delete/{idcompany}")
     public void deleteCompany(@PathVariable Long idcompany) {
-        companyService.deleteCompany(idcompany);
+        if(idcompany!=null) {
+            companyService.deleteCompany(idcompany);
+            System.out.println("Deleted");
+        }
+
     }
 
 }
